Moved models in proper app, and deleted the "old" version.
